+++ src/decorate.ts
