+++ src/extension.ts
@@ -22,7 +22,7 @@ let pendingDecorations: NodeJS.Timeout | undefined;
 
 import path = require('path');
 import * as vscode from 'vscode';
-import { DecoratorAndParser } from './parseAndDecorate';
+import { DecoratorAndParser } from './decorate';
 const snippetsJson = require("../snippets/snippets.json");
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
