+++ src/test/suite/extension.test.ts
@@ -18,7 +18,7 @@ import { EOL } from 'os';
 // You can import and use all API from the 'vscode' module
 // as well as import your extension to test it
 import * as vscode from 'vscode';
-import { placeholderDecoration, selectDecoration, pluralDecoration, DecoratorAndParser } from '../../parseAndDecorate';
+import { placeholderDecoration, selectDecoration, pluralDecoration, DecoratorAndParser } from '../../decorate';
 import { Parser } from '../../messageParser';
 
 const annotationNames = new Map<vscode.TextEditorDecorationType, string>([
